+++ apps/voting/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.2
 
 - Fix type error when closing wallet connect popup.
+- Fix deeplinking to wallets on iOS
+- Minor style changes
 
 ## 0.2.1